Move ParseAST into macroni::common
- Move `pasta::parse_ast()` function into a common include for macroni
  binaries to use.
- Add new cmake target, macroni::common
