+++ include/macroni/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(Common)
 add_subdirectory(Dialect)
 add_subdirectory(Conversion)
 add_subdirectory(Translation)
