+++ include/macroni/Common/CMakeLists.txt
@@ -1,0 +1,1 @@
+# Copyright (c) 2023-present, Trail of Bits, Inc.
