+++ lib/macroni/CMakeLists.txt
@@ -1,5 +1,6 @@
 # Copyright (c) 2023-present, Trail of Bits, Inc.
 
+add_subdirectory(Common)
 add_subdirectory(Dialect)
 add_subdirectory(Conversion)
 add_subdirectory(Translation)
