+++ lib/macroni/Common/ParseAST.cpp
@@ -1,4 +1,4 @@
-#include "ParseAST.hpp"
+#include <macroni/Common/ParseAST.hpp>
 
 namespace pasta {
     Result<AST, std::string> parse_ast(int argc, char **argv) {
